Search routes by osm_id
The file search.cc now supports the search of routes by the given id. It utilizes a new method of the routing_graph.h to find the locations of the osm-ids and then uses the normal method to find a route. You have the option to specify of which type the given osm-id is, as there is a node, an edge and an area with a given osm-id. If it is not specified it uses node > edge > area. With id-searcher you can test the new functionalities.