Handle requests with osm-ids
The parser of the http-requests can now handle requests where the start and/or the destination is given as an osm-id. The request should include a type of the start and destination. If there is no type present it handles it given the unknown type which checks node > area > edge.